Separated Google API from address logic
